requirements.txt
argon2-cffi-19.1.0
asn1crypto==0.24.0
bleach==3.0.2
blessings==1.7
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.5
curtsies==0.3.0
-e git+https://github.com/IQSS/dataverse-client-python.git@21b7837cb3f99634cb50d480e47874fd9269b4b9#egg=dataverse
enum34==1.1.6
greenlet==0.4.15
idna==2.7
ipaddress==1.0.22
lxml==4.2.5
miscreant==0.3.0
pycparser==2.19
Pygments==2.3.0
PyNaCl==1.3.0
requests==2.20.1
six==1.11.0
typing==3.6.6
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
